rearranged functional tests for vectors
