+++ data/csv/avg_cost_to_insert_into_arraylist.csv
@@ -1,300 +1,300 @@
 2,1
-3,1
-4,2
-5,3
-6,1
-7,2
-8,7
-9,5
+3,2
+4,1
+5,2
+6,4
+7,3
+8,3
+9,1
 10,0
-11,6
-12,10
-13,2
-14,8
-15,2
-16,8
+11,2
+12,7
+13,8
+14,7
+15,8
+16,7
 17,0
-18,5
-19,11
-20,18
-21,19
-22,20
-23,13
-24,14
-25,14
-26,5
-27,5
-28,5
+18,11
+19,8
+20,14
+21,12
+22,9
+23,21
+24,20
+25,16
+26,1
+27,16
+28,25
 29,0
-30,19
-31,1
-32,12
-33,30
-34,13
-35,19
-36,18
-37,29
-38,35
-39,26
-40,26
-41,1
-42,2
-43,39
-44,17
-45,18
-46,39
-47,3
-48,42
+30,13
+31,16
+32,2
+33,5
+34,19
+35,2
+36,13
+37,18
+38,4
+39,14
+40,3
+41,21
+42,34
+43,16
+44,28
+45,10
+46,41
+47,19
+48,13
 49,0
-50,41
-51,35
-52,2
-53,7
-54,23
-55,27
-56,14
-57,39
-58,11
-59,40
-60,12
-61,12
-62,54
+50,42
+51,13
+52,48
+53,6
+54,50
+55,49
+56,20
+57,2
+58,54
+59,54
+60,35
+61,14
+62,39
 63,46
-64,10
-65,49
-66,16
-67,30
-68,65
-69,26
-70,39
-71,39
-72,55
-73,36
-74,1
-75,15
-76,57
-77,72
-78,67
-79,35
-80,35
-81,64
-82,73
+64,32
+65,48
+66,45
+67,63
+68,54
+69,13
+70,13
+71,68
+72,70
+73,34
+74,44
+75,67
+76,24
+77,13
+78,38
+79,71
+80,15
+81,8
+82,68
 83,0
-84,65
-85,45
-86,15
-87,82
-88,14
-89,62
-90,4
-91,48
-92,14
-93,28
-94,67
-95,91
-96,15
-97,95
-98,69
-99,47
-100,96
-101,84
-102,50
-103,62
-104,78
-105,72
-106,103
-107,94
-108,22
-109,42
-110,84
-111,53
-112,91
-113,66
-114,46
-115,79
-116,108
-117,20
-118,33
-119,44
-120,64
-121,53
-122,85
-123,62
-124,119
-125,93
-126,97
-127,106
-128,13
-129,102
-130,3
-131,31
-132,61
-133,86
-134,99
-135,83
-136,62
-137,71
-138,13
-139,53
-140,94
+84,40
+85,64
+86,18
+87,16
+88,13
+89,22
+90,2
+91,74
+92,69
+93,85
+94,77
+95,49
+96,30
+97,88
+98,30
+99,17
+100,38
+101,2
+102,32
+103,31
+104,31
+105,55
+106,62
+107,97
+108,68
+109,89
+110,6
+111,63
+112,42
+113,90
+114,44
+115,108
+116,49
+117,76
+118,71
+119,6
+120,56
+121,93
+122,42
+123,60
+124,88
+125,101
+126,61
+127,95
+128,36
+129,93
+130,92
+131,25
+132,120
+133,62
+134,61
+135,112
+136,34
+137,92
+138,47
+139,49
+140,114
 141,0
-142,139
-143,39
-144,7
-145,46
-146,139
-147,83
-148,96
-149,26
-150,100
-151,15
-152,63
-153,1
-154,115
-155,45
-156,102
-157,72
-158,30
-159,103
-160,141
-161,107
-162,131
-163,86
-164,81
-165,119
-166,109
-167,53
-168,164
-169,166
-170,148
-171,118
-172,159
-173,167
-174,81
-175,4
-176,150
-177,4
-178,36
-179,156
-180,25
-181,74
-182,84
-183,64
-184,7
-185,159
-186,37
-187,168
-188,136
-189,137
-190,115
-191,2
-192,145
-193,167
-194,130
-195,168
-196,145
-197,42
-198,171
-199,188
-200,95
-201,44
-202,55
-203,9
-204,176
-205,76
-206,199
-207,64
-208,119
-209,37
-210,33
-211,43
-212,67
-213,193
-214,106
-215,79
-216,95
-217,63
-218,169
-219,67
-220,86
-221,115
-222,7
-223,106
-224,77
-225,207
-226,222
-227,120
-228,110
-229,14
-230,64
-231,125
-232,91
-233,221
-234,38
-235,160
-236,189
-237,118
-238,237
-239,142
+142,69
+143,86
+144,72
+145,13
+146,52
+147,112
+148,28
+149,75
+150,98
+151,61
+152,150
+153,18
+154,134
+155,57
+156,7
+157,53
+158,117
+159,5
+160,83
+161,67
+162,112
+163,35
+164,139
+165,62
+166,141
+167,102
+168,134
+169,137
+170,62
+171,168
+172,96
+173,6
+174,86
+175,98
+176,163
+177,106
+178,124
+179,93
+180,136
+181,137
+182,58
+183,14
+184,164
+185,120
+186,83
+187,117
+188,12
+189,107
+190,41
+191,111
+192,144
+193,190
+194,105
+195,2
+196,155
+197,190
+198,96
+199,136
+200,80
+201,82
+202,93
+203,22
+204,42
+205,75
+206,97
+207,153
+208,54
+209,86
+210,28
+211,49
+212,120
+213,79
+214,71
+215,159
+216,50
+217,188
+218,149
+219,185
+220,81
+221,71
+222,191
+223,152
+224,178
+225,170
+226,93
+227,41
+228,179
+229,30
+230,146
+231,165
+232,29
+233,214
+234,179
+235,111
+236,149
+237,67
+238,132
+239,61
 240,0
-241,116
-242,59
-243,126
-244,99
-245,232
-246,175
-247,196
-248,34
-249,18
-250,24
-251,223
-252,39
-253,199
-254,134
-255,133
-256,179
-257,62
-258,166
-259,108
+241,189
+242,216
+243,4
+244,157
+245,225
+246,71
+247,213
+248,47
+249,212
+250,167
+251,225
+252,164
+253,36
+254,70
+255,122
+256,104
+257,194
+258,69
+259,209
 260,245
-261,219
-262,104
-263,127
-264,260
-265,115
-266,38
-267,22
-268,195
-269,191
-270,153
-271,144
-272,65
-273,172
-274,71
-275,86
-276,92
-277,33
-278,162
-279,92
-280,51
-281,46
-282,108
-283,265
-284,268
-285,95
-286,59
-287,97
-288,133
-289,271
-290,33
-291,255
-292,133
-293,112
-294,179
-295,196
-296,7
-297,231
-298,14
-299,187
-300,44
-301,118
+261,12
+262,34
+263,19
+264,109
+265,29
+266,176
+267,223
+268,126
+269,126
+270,198
+271,211
+272,265
+273,125
+274,152
+275,6
+276,210
+277,243
+278,77
+279,8
+280,82
+281,168
+282,18
+283,256
+284,161
+285,24
+286,44
+287,118
+288,22
+289,32
+290,184
+291,271
+292,53
+293,118
+294,55
+295,139
+296,135
+297,267
+298,134
+299,92
+300,39
+301,54
