+++ data/csv/avg_cost_to_remove_from_arraylist.csv
@@ -1,300 +1,300 @@
-300,295
-299,22
-298,251
-297,67
-296,36
-295,229
-294,182
-293,273
-292,176
-291,175
-290,80
-289,79
-288,127
-287,33
-286,107
-285,285
-284,20
-283,17
-282,163
-281,80
-280,107
-279,13
-278,95
-277,155
-276,92
-275,123
-274,4
-273,163
-272,29
-271,130
-270,175
-269,40
-268,231
-267,16
-266,14
-265,177
-264,82
-263,148
-262,69
-261,121
-260,48
-259,169
-258,60
-257,100
-256,72
-255,89
-254,248
-253,240
-252,151
-251,109
-250,206
-249,40
-248,194
-247,140
-246,5
-245,87
-244,155
-243,204
-242,88
-241,112
+300,155
+299,28
+298,167
+297,98
+296,284
+295,73
+294,89
+293,200
+292,224
+291,164
+290,272
+289,57
+288,73
+287,109
+286,100
+285,140
+284,94
+283,265
+282,178
+281,92
+280,153
+279,175
+278,64
+277,131
+276,118
+275,186
+274,203
+273,43
+272,127
+271,24
+270,30
+269,91
+268,224
+267,220
+266,156
+265,39
+264,93
+263,6
+262,100
+261,111
+260,111
+259,259
+258,245
+257,149
+256,246
+255,72
+254,47
+253,75
+252,47
+251,188
+250,37
+249,213
+248,128
+247,35
+246,72
+245,157
+244,190
+243,44
+242,87
+241,127
 240,123
-239,136
-238,207
-237,2
-236,229
-235,107
-234,51
-233,86
-232,47
-231,118
-230,61
-229,23
-228,62
-227,174
-226,110
-225,1
-224,222
-223,66
-222,28
-221,202
-220,21
-219,21
-218,77
-217,101
-216,139
-215,139
-214,34
-213,122
-212,191
-211,151
-210,149
-209,160
-208,8
-207,186
-206,161
-205,23
-204,123
-203,13
-202,106
-201,161
-200,1
-199,187
-198,31
-197,6
-196,74
-195,120
-194,70
-193,65
-192,30
-191,27
-190,166
-189,33
-188,180
-187,118
-186,161
-185,123
-184,122
-183,116
-182,68
-181,21
-180,99
-179,121
-178,34
-177,31
-176,69
-175,165
-174,33
-173,12
-172,162
-171,137
-170,33
-169,142
-168,52
-167,76
-166,18
-165,159
-164,19
-163,95
-162,31
-161,139
-160,122
-159,28
-158,69
-157,149
-156,112
-155,99
-154,83
-153,139
-152,2
-151,48
-150,3
-149,31
-148,45
-147,77
-146,113
-145,40
-144,80
-143,48
-142,12
-141,129
-140,102
-139,115
-138,28
-137,85
-136,22
-135,88
-134,31
-133,48
-132,17
-131,49
-130,129
-129,3
-128,53
-127,18
-126,107
-125,82
-124,50
-123,22
-122,73
-121,101
-120,54
-119,35
-118,86
-117,95
-116,5
-115,99
-114,15
-113,39
-112,12
-111,9
-110,35
-109,21
-108,22
-107,29
-106,49
-105,65
-104,2
-103,9
-102,96
-101,13
-100,2
-99,80
-98,66
-97,24
-96,31
-95,93
-94,8
-93,24
-92,83
-91,57
-90,15
-89,61
+239,162
+238,107
+237,136
+236,42
+235,228
+234,36
+233,115
+232,174
+231,112
+230,216
+229,152
+228,223
+227,25
+226,98
+225,189
+224,72
+223,180
+222,190
+221,126
+220,167
+219,83
+218,12
+217,36
+216,47
+215,85
+214,165
+213,113
+212,135
+211,47
+210,65
+209,39
+208,158
+207,193
+206,124
+205,24
+204,170
+203,187
+202,101
+201,156
+200,102
+199,130
+198,88
+197,179
+196,62
+195,35
+194,164
+193,6
+192,6
+191,137
+190,137
+189,9
+188,1
+187,146
+186,147
+185,163
+184,16
+183,52
+182,61
+181,5
+180,88
+179,76
+178,145
+177,144
+176,141
+175,88
+174,64
+173,70
+172,49
+171,97
+170,77
+169,61
+168,145
+167,25
+166,145
+165,26
+164,131
+163,139
+162,61
+161,155
+160,94
+159,155
+158,111
+157,117
+156,145
+155,136
+154,52
+153,33
+152,73
+151,119
+150,84
+149,121
+148,11
+147,137
+146,3
+145,137
+144,9
+143,127
+142,36
+141,73
+140,82
+139,91
+138,81
+137,33
+136,107
+135,121
+134,129
+133,84
+132,109
+131,95
+130,81
+129,108
+128,125
+127,73
+126,21
+125,38
+124,64
+123,75
+122,11
+121,36
+120,82
+119,92
+118,75
+117,100
+116,70
+115,101
+114,98
+113,86
+112,111
+111,100
+110,4
+109,63
+108,74
+107,25
+106,58
+105,101
+104,78
+103,82
+102,73
+101,26
+100,19
+99,92
+98,22
+97,27
+96,91
+95,82
+94,51
+93,15
+92,26
+91,67
+90,67
+89,85
 88,70
-87,70
-86,27
-85,75
-84,40
-83,58
-82,11
-81,77
-80,29
-79,68
-78,44
-77,31
-76,55
-75,45
-74,47
-73,63
-72,2
-71,3
-70,35
-69,20
-68,23
-67,60
-66,49
-65,19
-64,7
-63,48
-62,10
-61,31
-60,11
-59,18
-58,56
-57,22
-56,48
-55,9
-54,17
-53,35
-52,19
-51,8
-50,15
-49,2
-48,26
-47,17
-46,44
-45,19
-44,38
-43,4
-42,3
-41,9
-40,35
-39,17
-38,28
-37,26
-36,9
-35,27
-34,18
-33,7
-32,27
-31,16
-30,22
-29,13
-28,3
-27,18
-26,4
-25,14
+87,23
+86,56
+85,74
+84,28
+83,6
+82,2
+81,75
+80,41
+79,66
+78,58
+77,20
+76,54
+75,60
+74,35
+73,15
+72,29
+71,33
+70,30
+69,11
+68,65
+67,11
+66,51
+65,58
+64,11
+63,7
+62,16
+61,1
+60,34
+59,14
+58,41
+57,20
+56,33
+55,5
+54,20
+53,7
+52,8
+51,30
+50,32
+49,21
+48,20
+47,10
+46,20
+45,7
+44,1
+43,34
+42,37
+41,32
+40,6
+39,33
+38,34
+37,19
+36,23
+35,17
+34,21
+33,1
+32,22
+31,17
+30,12
+29,12
+28,23
+27,10
+26,3
+25,16
 24,7
-23,17
-22,21
-21,7
-20,12
-19,6
-18,11
-17,2
-16,5
-15,8
-14,7
-13,2
-12,9
-11,11
-10,1
-9,2
-8,2
-7,1
-6,3
-5,4
-4,4
-3,3
-2,1
+23,9
+22,14
+21,18
+20,9
+19,8
+18,14
+17,8
+16,1
+15,12
+14,4
+13,6
+12,8
+11,8
+10,6
+9,4
+8,3
+7,5
+6,5
+5,3
+4,3
+3,1
+2,2
 1,1
