+++ data/csv/levels_in_avl_tree_vs_n_random_keys.csv
@@ -8,7 +8,7 @@
 8,4
 9,4
 10,4
-11,5
+11,4
 12,5
 13,5
 14,5
@@ -19,7 +19,7 @@
 19,5
 20,5
 21,5
-22,5
+22,6
 23,6
 24,6
 25,6
@@ -33,17 +33,17 @@
 33,6
 34,6
 35,6
-36,6
-37,6
-38,6
-39,6
-40,6
-41,6
-42,6
-43,6
-44,6
-45,6
-46,6
+36,7
+37,7
+38,7
+39,7
+40,7
+41,7
+42,7
+43,7
+44,7
+45,7
+46,7
 47,7
 48,7
 49,7
@@ -53,7 +53,7 @@
 53,7
 54,7
 55,7
-56,7
+56,8
 57,7
 58,7
 59,7
@@ -62,8 +62,8 @@
 62,7
 63,7
 64,7
-65,8
-66,8
+65,7
+66,7
 67,7
 68,7
 69,7
