+++ data/csv/levels_in_bst_tree_vs_n_random_keys.csv
@@ -3,63 +3,64 @@
 3,3
 4,4
 5,4
-6,5
-7,6
-8,6
-9,6
-10,7
-11,7
-12,7
-13,7
+6,4
+7,4
+8,4
+9,5
+10,5
+11,5
+12,6
+13,6
 14,7
 15,7
 16,7
-17,7
-18,7
-19,7
-20,7
-21,8
-22,8
-23,8
+17,8
+18,9
+19,9
+20,9
+21,9
+22,9
+23,9
 24,9
 25,9
 26,9
 27,9
 28,9
 29,9
-30,9
-31,9
-32,9
-33,9
+30,10
+31,10
+32,10
+33,10
 34,10
-35,11
+35,10
 36,11
-37,12
-38,12
-39,12
-40,12
-41,12
-42,12
-43,12
-44,12
-45,12
-46,12
-47,13
-48,13
-49,13
-50,13
-51,13
-52,13
-53,13
-54,13
-55,13
-56,13
-57,13
-58,13
-59,13
-60,13
-61,13
+37,11
+38,11
+39,11
+40,11
+41,11
+42,11
+43,11
+44,11
+45,11
+46,11
+47,11
+48,11
+49,11
+50,11
+51,12
+52,12
+53,12
+54,12
+55,12
+56,12
+57,12
+58,12
+59,12
+60,12
+61,12
 62,13
+63,13
 63,13
 64,13
 65,13
@@ -90,11 +91,10 @@
 90,13
 91,13
 92,13
-93,14
-94,14
-95,14
-96,14
-97,14
-98,14
-99,14
-100,14
+93,13
+94,13
+95,13
+96,13
+97,13
+98,13
+99,13
