+++ data/csv/shifts_required_when_inserting_rand_index.csv
@@ -1,300 +1,300 @@
 2,1
 3,2
-4,5
-5,9
-6,13
+4,4
+5,5
+6,9
 7,14
-8,17
-9,25
-10,25
-11,29
-12,37
-13,46
-14,50
-15,57
+8,15
+9,23
+10,23
+11,28
+12,33
+13,44
+14,52
+15,56
 16,66
 17,66
-18,67
-19,82
-20,97
-21,103
-22,123
-23,140
-24,152
-25,163
-26,177
-27,191
-28,208
-29,208
-30,228
-31,235
-32,241
-33,252
-34,266
-35,289
-36,307
-37,316
-38,341
-39,374
-40,403
-41,431
-42,461
-43,497
-44,526
-45,564
+18,71
+19,88
+20,107
+21,121
+22,137
+23,150
+24,166
+25,169
+26,191
+27,202
+28,205
+29,205
+30,229
+31,231
+32,256
+33,279
+34,306
+35,319
+36,346
+37,351
+38,375
+39,382
+40,389
+41,414
+42,443
+43,480
+44,521
+45,543
 46,587
-47,622
-48,659
-49,659
-50,684
-51,714
-52,739
-53,767
-54,793
-55,834
-56,865
-57,901
-58,954
-59,995
-60,1044
-61,1052
-62,1113
-63,1130
-64,1167
-65,1194
-66,1245
-67,1276
-68,1314
-69,1370
-70,1384
-71,1446
-72,1462
-73,1508
-74,1539
-75,1594
-76,1667
-77,1678
-78,1708
-79,1722
-80,1755
-81,1834
-82,1874
-83,1874
-84,1948
-85,1969
-86,2041
-87,2110
-88,2151
-89,2170
-90,2248
-91,2305
-92,2356
-93,2387
-94,2428
-95,2429
-96,2436
-97,2523
-98,2611
-99,2678
-100,2693
-101,2701
-102,2796
-103,2807
-104,2892
-105,2940
-106,3033
-107,3137
-108,3161
-109,3269
-110,3284
-111,3375
-112,3377
-113,3465
-114,3529
-115,3562
-116,3611
-117,3613
-118,3683
-119,3794
-120,3807
-121,3892
-122,4008
-123,4037
-124,4062
-125,4162
-126,4216
-127,4309
-128,4347
-129,4465
-130,4557
-131,4645
-132,4769
-133,4885
-134,4917
-135,4935
-136,5014
-137,5085
-138,5113
-139,5223
-140,5301
-141,5301
-142,5332
-143,5467
-144,5563
-145,5600
-146,5659
-147,5686
-148,5787
-149,5820
-150,5839
-151,5844
-152,5881
-153,6033
-154,6034
-155,6116
-156,6172
-157,6285
-158,6358
-159,6393
-160,6430
-161,6525
-162,6563
-163,6678
-164,6698
-165,6735
-166,6864
-167,7014
-168,7060
-169,7174
-170,7270
-171,7371
-172,7376
-173,7506
-174,7646
-175,7685
-176,7713
-177,7889
-178,8051
-179,8208
-180,8249
-181,8310
-182,8372
-183,8406
-184,8505
-185,8640
-186,8770
-187,8779
-188,8797
-189,8902
-190,8979
-191,9134
-192,9224
-193,9405
-194,9555
-195,9671
-196,9709
-197,9854
-198,10009
-199,10059
-200,10160
-201,10304
-202,10368
-203,10409
-204,10466
-205,10589
-206,10640
-207,10671
-208,10851
-209,10974
-210,11060
-211,11070
-212,11259
-213,11388
-214,11579
-215,11669
-216,11841
-217,11858
-218,11961
-219,12089
-220,12218
-221,12287
-222,12341
-223,12503
-224,12669
-225,12888
-226,12944
-227,13158
-228,13356
-229,13557
-230,13581
-231,13689
-232,13880
-233,14024
-234,14254
-235,14368
-236,14370
-237,14587
-238,14717
-239,14881
-240,14881
-241,15028
-242,15034
-243,15156
-244,15166
-245,15201
-246,15259
-247,15371
-248,15403
-249,15465
-250,15643
-251,15718
-252,15812
-253,16011
-254,16255
-255,16384
-256,16495
-257,16564
-258,16584
-259,16807
-260,16999
-261,17158
-262,17387
-263,17536
-264,17641
-265,17760
-266,17834
-267,17944
-268,18038
-269,18279
-270,18305
-271,18552
-272,18643
-273,18812
-274,18892
-275,19014
-276,19021
-277,19105
-278,19261
-279,19305
-280,19319
-281,19413
-282,19591
-283,19853
-284,19946
-285,20200
-286,20382
-287,20429
-288,20520
-289,20693
-290,20948
-291,21205
-292,21286
-293,21499
-294,21701
-295,21702
-296,21715
-297,21716
-298,21882
-299,22093
-300,22099
-301,22209
+47,616
+48,648
+49,648
+50,664
+51,666
+52,686
+53,695
+54,697
+55,720
+56,773
+57,783
+58,824
+59,829
+60,855
+61,911
+62,928
+63,971
+64,1015
+65,1041
+66,1070
+67,1076
+68,1089
+69,1115
+70,1118
+71,1168
+72,1206
+73,1268
+74,1332
+75,1397
+76,1465
+77,1474
+78,1505
+79,1558
+80,1598
+81,1660
+82,1726
+83,1726
+84,1756
+85,1787
+86,1794
+87,1840
+88,1892
+89,1942
+90,2022
+91,2042
+92,2091
+93,2128
+94,2154
+95,2244
+96,2260
+97,2304
+98,2365
+99,2402
+100,2477
+101,2562
+102,2603
+103,2649
+104,2732
+105,2752
+106,2783
+107,2794
+108,2814
+109,2821
+110,2878
+111,2946
+112,2949
+113,3052
+114,3060
+115,3108
+116,3220
+117,3304
+118,3379
+119,3396
+120,3441
+121,3499
+122,3524
+123,3613
+124,3635
+125,3651
+126,3724
+127,3758
+128,3869
+129,3944
+130,3958
+131,4005
+132,4079
+133,4203
+134,4295
+135,4304
+136,4387
+137,4497
+138,4632
+139,4737
+140,4771
+141,4771
+142,4881
+143,4901
+144,4984
+145,5102
+146,5182
+147,5257
+148,5378
+149,5424
+150,5549
+151,5631
+152,5634
+153,5649
+154,5689
+155,5774
+156,5836
+157,5871
+158,5925
+159,5928
+160,5962
+161,6068
+162,6212
+163,6269
+164,6297
+165,6343
+166,6422
+167,6448
+168,6508
+169,6573
+170,6697
+171,6822
+172,6893
+173,6931
+174,6933
+175,7071
+176,7163
+177,7247
+178,7365
+179,7485
+180,7537
+181,7558
+182,7736
+183,7745
+184,7909
+185,8045
+186,8063
+187,8176
+188,8327
+189,8335
+190,8491
+191,8561
+192,8752
+193,8779
+194,8825
+195,8833
+196,8969
+197,9076
+198,9222
+199,9263
+200,9399
+201,9545
+202,9613
+203,9656
+204,9721
+205,9873
+206,10006
+207,10129
+208,10231
+209,10370
+210,10390
+211,10517
+212,10544
+213,10720
+214,10773
+215,10987
+216,11044
+217,11150
+218,11323
+219,11512
+220,11693
+221,11723
+222,11821
+223,11955
+224,12143
+225,12210
+226,12426
+227,12543
+228,12613
+229,12806
+230,12808
+231,12885
+232,12938
+233,13168
+234,13275
+235,13473
+236,13673
+237,13715
+238,13913
+239,14048
+240,14048
+241,14183
+242,14214
+243,14304
+244,14425
+245,14551
+246,14615
+247,14732
+248,14931
+249,14969
+250,15140
+251,15302
+252,15305
+253,15431
+254,15675
+255,15909
+256,15983
+257,16169
+258,16404
+259,16617
+260,16785
+261,17007
+262,17184
+263,17246
+264,17486
+265,17695
+266,17917
+267,18161
+268,18343
+269,18542
+270,18636
+271,18808
+272,18931
+273,19011
+274,19138
+275,19319
+276,19445
+277,19624
+278,19818
+279,19929
+280,20196
+281,20464
+282,20737
+283,20803
+284,20961
+285,21122
+286,21181
+287,21276
+288,21408
+289,21441
+290,21700
+291,21726
+292,21888
+293,22094
+294,22258
+295,22402
+296,22639
+297,22825
+298,22889
+299,22947
+300,23185
+301,23410
