+++ src/main/java/structures/sets/DisjointSet.java
@@ -1,0 +1,4 @@
+package structures.sets;
+
+public class DisjointSet {
+}
