+++ src/main/java/structures/vectors/ArrayQueue.java
@@ -8,6 +8,8 @@ import java.util.Collection;
  * Implementation of Queue using
  * an internal array to hold data elements.
  *
+ * TODO - https://www.geeksforgeeks.org/queue-set-1introduction-and-array-implementation/
+ *
  * @author Jabari Dash
  * @param <E> Generic type
  */
